requirements.txt
BeautifulSoup4
python-dotenv
twilio